% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeros.R
\name{mil}
\alias{mil}
\title{Numero em milhares}
\usage{
mil(x, dec = 0)
}
\arguments{
\item{x}{Vetor numerico.}

\item{dec}{Casas decimais depois de dividir por mil.}
}
\value{
Vetor de texto.
}
\description{
Para eixos e rotulos onde a unidade inteira polui. \code{mil(137063)} devolve
"137", nao "137.063".
}
\examples{
mil(137063)  # "137"
}
