% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeros.R
\name{num_br}
\alias{num_br}
\title{Numero em pt-BR}
\usage{
num_br(x, dec = 0)
}
\arguments{
\item{x}{Vetor numerico.}

\item{dec}{Casas decimais.}
}
\value{
Vetor de texto do mesmo comprimento de \code{x}.
}
\description{
Milhar com ponto, decimal com virgula. Serve tanto dentro de uma tabela
quanto colado no texto de um artigo.
}
\examples{
num_br(1234567)          # "1.234.567"
num_br(33.126, dec = 2)  # "33,13"
}
