% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeros.R
\name{pct_br}
\alias{pct_br}
\title{Percentual em pt-BR}
\usage{
pct_br(x, dec = 1)
}
\arguments{
\item{x}{Vetor numerico ja em escala de 0 a 100 (nao de 0 a 1).}

\item{dec}{Casas decimais.}
}
\value{
Vetor de texto, com o sinal de porcento colado.
}
\description{
Percentual em pt-BR
}
\examples{
pct_br(33.126)  # "33,1\%"
}
