% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabelas.R
\name{fmt_pct_br}
\alias{fmt_pct_br}
\title{Formata colunas de percentual em pt-BR}
\usage{
fmt_pct_br(g, colunas, decimais = 1)
}
\arguments{
\item{g}{Objeto \code{gt_tbl}.}

\item{colunas}{Colunas a formatar (tidyselect).}

\item{decimais}{Casas decimais.}
}
\value{
O \code{gt_tbl} com as colunas formatadas como percentual.
}
\description{
Formata colunas de percentual em pt-BR
}
